docs: add note about incompatibility with React Native 0.80
Added a note in the README to inform users that the @pushpendersingh/react-native-scanner package is currently not compatible with React Native 0.80 due to recent architectural changes. 

Encouraged users to stay on React Native 0.79 or below for now and invited contributors to open a pull request if they’d like to help make the library compatible with RN 0.80.

This update is intended to reduce confusion and set expectations around version support.